requirements.txt
blinker==1.9.0
click==8.1.8
cloudinary==1.32.0
colorama==0.4.6
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
reportlab==3.6.13
